+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/MainTabView.swift
@@ -32,7 +32,7 @@ struct MainTabView: View {
             GuidelinesList()
                 .tabItem {
                     Label("Guidelines", image: "Guideline-DNA_32")
-                }.navigationBarColor(tintColor: UIColor(ODS.coreOrange))
+                }
             ComponentsCardsList()
                 .tabItem {
                     Label("Components", image: "component-atom_32")
@@ -45,13 +45,15 @@ struct MainTabView: View {
                 .tabItem {
                     Label("About", image: "info_32")
                 }
-            Text("Search")
-                .font(ODSFontStyle.largeTitle.font())
-                .tabItem {
-                    Label("Search", image: "Search_32")
-                }
+//            Text("Search")
+//                .font(ODSFontStyle.largeTitle.font())
+//                .tabItem {
+//                    Label("Search", image: "Search_32")
+//                }
         }
-        .tabBarStyle(itemColor: Color.primary)
+        .navigationBarColor(tintColor: ODSColor.coreOrange.color.uiColor ?? UIColor.orange,
+                            backgroundColor: ODSColor.componentBackground.color.uiColor ?? UIColor.systemBackground)
+        .tabBarStyle(backgroundColor: ODSColor.componentBackground.color, itemColor: Color.primary)
     }
 }