+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/modules/CardViewDemo.swift
@@ -59,10 +59,10 @@ struct CardViewDemoGrid: View {
                 }
             }.padding([.leading, .trailing])
             Spacer().frame(height: 15)
-
-        }.navigationTitle("Page title")
-            .navigationViewStyle(.stack)
-            .background(Color(uiColor: .systemGray5))
+        }
+        .navigationTitle("Page title")
+        .navigationViewStyle(.stack)
+        .background(ODSColor.primaryBackground.color)
     }
 }