+++ OrangeDesignSystemDemo/Pages/TabBarPage.swift
@@ -30,25 +30,29 @@ struct TabBarPage: View {
 
     var body: some View {
         ScrollView {
-            Image("empty")
+            Image("Tab bar")
                 .resizable()
                 .aspectRatio(contentMode: .fill)
             VStack(alignment: .leading, spacing: 20) {
                 ComponentDescription(text: "A tab bar is a constantly available element which creates an overall navigation for users' experience.")
                 VariantsTitle()
-                TabBarVariant()
+                TabBarVariant(nbItems: 2)
+                TabBarVariant(nbItems: 3)
+                TabBarVariant(nbItems: 4)
+                TabBarVariant(nbItems: 5)
                 Spacer().frame(height: 10)
             }.padding(EdgeInsets(top: 0, leading: 15, bottom: 5, trailing: 15))
-        }.navigationTitle("Tab bar")
-            .navigationViewStyle(.stack)
-            .background(Color(uiColor: .systemGray6))
+        }.background(ODSColor.primaryBackground.color)
     }
 }
 
 struct TabBarVariant: View {
 
+    var nbItems: Int
+
     var body: some View {
-        Image("empty")
+        Text("\(nbItems) entries").odsFont(.title2)
+        Image("TabBar-\(nbItems)-items")
             .resizable()
             .aspectRatio(contentMode: .fill)
     }